Tesla Model S: Use 32bit arithmetic for data decoding

+++ vehicle/OVMS.V3/components/vehicle_teslamodels/src/vehicle_teslamodels.cpp
@@ -90,12 +90,12 @@ void OvmsVehicleTeslaModelS::IncomingFrameCan1(CAN_frame_t* p_frame)
       }
     case 0x256: // Speed
       {
-      StandardMetrics.ms_v_pos_speed->SetValue( (((d[3]&0xf0)<<8) + d[2])/10, Mph );
+      StandardMetrics.ms_v_pos_speed->SetValue( ((((int)d[3]&0xf0)<<8) + (int)d[2])/10, Mph );
       break;
       }
     case 0x302: // SOC
       {
-      StandardMetrics.ms_v_bat_soc->SetValue( ((d[1]>>2) + ((d[2] & 0xf0)<<6))/10 );
+      StandardMetrics.ms_v_bat_soc->SetValue( (((int)d[1]>>2) + (((int)d[2] & 0xf0)<<6))/10 );
       break;
       }
     case 0x398: // Country
